test: move util function in utils file
