+++ backend/test_utils.py
@@ -1,3 +1,18 @@
+from requests import put
+from requests import post
+from requests import get
+
+from json import loads
+
+
+class bcolors:
+    OKGREEN = "\033[92m"
+    ENDC = "\033[0m"
+
+
+URL = "http://localhost:5001"
+
+
 def print_error_post(
     url, path, json, code_expected, code_received, content_expected, content_received
 ):
@@ -23,3 +38,115 @@ def print_error_get(
     print(f"code expected: {code_expected}")
     print(f"code received: {code_received}")
     print("----")
+
+
+def check_login_token(path, json):
+    headers = {"content-type": "application/json"}
+
+    response = post(URL + path, json=json, headers=headers)
+
+    if response.status_code != 200:
+        print(f"error: status code {response.status_code}")
+
+    json = loads(response.content)
+
+    if "access_token" not in json:
+        print("error: access token not in response")
+
+    print(bcolors.OKGREEN + "success: got access token" + bcolors.ENDC)
+
+    global access_token
+
+    access_token = json["access_token"]
+
+
+def check_415(path):
+    headers = {"content-type": "application/not_json"}
+
+    json = {}
+
+    response = post(URL + path, json=json, headers=headers)
+
+    if response.status_code != 415:
+        print(f"error: status code {response.status_code}")
+
+    print(bcolors.OKGREEN + "success: " + bcolors.ENDC + path + " 415 error")
+
+
+def check_api_get(path, status, content):
+    response = get(URL + path)
+
+    if response.content != content:
+        print("--- bad content ---")
+        print_error_get(
+            URL, path, status, response.status_code, content, response.content
+        )
+        return
+    if response.status_code != status:
+        print("--- bad http code ---")
+        print_error_get(
+            URL, path, status, response.status_code, content, response.content
+        )
+
+    print(bcolors.OKGREEN + "success: " + bcolors.ENDC + path + str(content))
+
+
+def check_api_put(path, status, json, content):
+    headers = {"content-type": "application/json"}
+
+    response = put(URL + path, json=json, headers=headers)
+
+    if response.content != content:
+        print("--- bad content ---")
+        print_error_post(
+            URL, path, json, status, response.status_code, content, response.content
+        )
+        return
+    if response.status_code != status:
+        print("--- bad http code ---")
+        print_error_post(
+            URL, path, json, status, response.status_code, content, response.content
+        )
+        return
+
+    print(bcolors.OKGREEN + "success: " + bcolors.ENDC + path + " " + str(content))
+
+
+def check_api_post(path, status, json, content):
+    headers = {"content-type": "application/json"}
+
+    response = post(URL + path, json=json, headers=headers)
+
+    if response.content != content:
+        print("--- bad content ---")
+        print_error_post(
+            URL, path, json, status, response.status_code, content, response.content
+        )
+        return
+    if response.status_code != status:
+        print("--- bad http code ---")
+        print_error_post(
+            URL, path, json, status, response.status_code, content, response.content
+        )
+        return
+
+    print(bcolors.OKGREEN + "success: " + bcolors.ENDC + path + " " + str(content))
+
+
+def check_put_token(path, json, content):
+    headers = {"Authorization": f"Bearer {access_token}"}
+
+    response = put(URL + path, headers=headers, json=json)
+
+    if response.content != content:
+        print(f"error: content {response.content}")
+        return
+
+    print(
+        bcolors.OKGREEN
+        + "success: "
+        + bcolors.ENDC
+        + path
+        + " "
+        + str(response.content)
+    )
